requirements.txt
einops
equinox
h5py
hydra-core
hydra-submitit-launcher
imageio
ipdb
jaxtyping
matplotlib
moviepy
optax
pandas
scikit-learn
seaborn
sigfig
tensorflow
tensorflow-datasets
tqdm
wandb